requirements.txt
# Core UI Framework
kivy==2.2.0
kivymd==1.1.1

# Image Processing (minimal)
pillow==10.0.1

# Barcode Scanning
pyzbar==0.1.9

# Data Processing
pandas==2.1.1
openpyxl==3.1.2

# Claude AI API
anthropic==0.7.8

# Mobile Features
plyer==2.1.0